Kernel/Net: Introduce proper pointers for payload calls
Unifying with other structures, this introduces m_payload for all
relevant structures, except the TCP header (which has a different
concept of payload than the rest).

This allows us to return a pointer directly instead of doing pointer
arithmetic to point to the end of the structure.

+++ Kernel/Net/EthernetFrameHeader.h
@@ -30,7 +30,7 @@ private:
     MACAddress m_destination;
     MACAddress m_source;
     NetworkOrdered<u16> m_ether_type;
-    u32 m_payload[0];
+    u8 m_payload[0];
 };
 
 static_assert(sizeof(EthernetFrameHeader) == 14);
